requirements.txt
Django>=3.2
djangorestframework
psycopg2-binary
dj-database-url
python-dotenv
drf-yasg
Faker
dj-database-url
datetime
